+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/CosmosVectorEmbeddingPolicy.java
@@ -14,10 +14,9 @@ import java.util.List;
 /**
  * Vector Embedding Policy
  */
-public final class VectorEmbeddingPolicy {
+public final class CosmosVectorEmbeddingPolicy {
 
     private JsonSerializable jsonSerializable;
-
     /**
      * Paths for embeddings along with path-specific settings for the item.
      */
@@ -29,16 +28,10 @@ public final class VectorEmbeddingPolicy {
      *
      * @param cosmosVectorEmbeddings list of path for embeddings along with path-specific settings for the item.
      */
-    public VectorEmbeddingPolicy(List<CosmosVectorEmbedding> cosmosVectorEmbeddings) {
+    public CosmosVectorEmbeddingPolicy(List<CosmosVectorEmbedding> cosmosVectorEmbeddings) {
         validateEmbeddings(cosmosVectorEmbeddings);
-        this.cosmosVectorEmbeddings = cosmosVectorEmbeddings;
-    }
-
-    /**
-     * Constructor.
-     */
-    public VectorEmbeddingPolicy() {
         this.jsonSerializable = new JsonSerializable();
+        this.setCosmosVectorEmbeddings(cosmosVectorEmbeddings);
     }
 
     private static void validateEmbeddings(List<CosmosVectorEmbedding> cosmosVectorEmbeddings) {
@@ -48,14 +41,14 @@ public final class VectorEmbeddingPolicy {
             }
             validateEmbeddingPath(embedding.getPath());
             validateEmbeddingDimensions(embedding.getDimensions());
-            validateEmbeddingVectorDataType(embedding.getCosmosVectorDataType());
-            validateEmbeddingDistanceFunction(embedding.getCosmosVectorDistanceFunction());
+            validateEmbeddingVectorDataType(embedding.getVectorDataType());
+            validateEmbeddingDistanceFunction(embedding.getVectorDistanceFunction());
         });
     }
 
     private static void validateEmbeddingPath(String path) {
         if (StringUtils.isEmpty(path)) {
-            throw new IllegalArgumentException("embedding path is empty");
+            throw new NullPointerException("embedding path is empty");
         }
 
         if (path.charAt(0) != '/' || path.lastIndexOf('/') != 0) {
@@ -65,7 +58,7 @@ public final class VectorEmbeddingPolicy {
 
     private static void validateEmbeddingDimensions(Long dimensions) {
         if (dimensions == null) {
-            throw new IllegalArgumentException("Dimensions for the embedding cannot be null " +
+            throw new NullPointerException("Dimensions for the embedding cannot be null " +
                 "for the vector embedding policy");
         }
         if (dimensions < 1) {
@@ -74,16 +67,16 @@ public final class VectorEmbeddingPolicy {
         }
     }
 
-    private static void validateEmbeddingVectorDataType(String value) {
+    private static void validateEmbeddingVectorDataType(CosmosVectorDataType value) {
         if (Arrays.stream(CosmosVectorDataType.values()).noneMatch(vectorDataType ->
-            vectorDataType.toString().equals(value))) {
+            vectorDataType.equals(value))) {
             throw new IllegalArgumentException("Invalid vector data type for the vector embedding policy.");
         }
     }
 
-    private static void validateEmbeddingDistanceFunction(String value) {
+    private static void validateEmbeddingDistanceFunction(CosmosVectorDistanceFunction value) {
         if (Arrays.stream(CosmosVectorDistanceFunction.values()).noneMatch(distanceFunction ->
-            distanceFunction.toString().equals(value))) {
+            distanceFunction.equals(value))) {
             throw new IllegalArgumentException("Invalid distance function for the vector embedding policy.");
         }
     }
@@ -93,7 +86,17 @@ public final class VectorEmbeddingPolicy {
      *
      * @return the paths for embeddings along with path-specific settings for the item.
      */
-    public List<CosmosVectorEmbedding> getCosmosVectorEmbeddings() {
+    public List<CosmosVectorEmbedding> getVectorEmbeddings() {
         return this.cosmosVectorEmbeddings;
     }
+
+    /**
+     * Sets the paths for embeddings along with path-specific settings for the item.
+     *
+     * @param cosmosVectorEmbeddings paths for embeddings along with path-specific settings for the item.
+     */
+    public void setCosmosVectorEmbeddings(List<CosmosVectorEmbedding> cosmosVectorEmbeddings) {
+        this.cosmosVectorEmbeddings = cosmosVectorEmbeddings;
+    }
+
 }